GetLocationService Bean 분리
